test(writer): consolidate drive-file-meta and search tests into one file
Both PR #622 (drive file meta cache) and PR #520 (search metadata
leakage) added test_general.py for suite/writer/api/general.py at two
different paths. Merge them into suite/writer/api/tests/test_general.py,
matching the tests/ package convention used by drive/api, sheets, and
calendar, and add the __init__.py that convention expects.
